+++ WoWDeveloperAssistant/DB2/Structures/Criteria.cs
@@ -6,12 +6,12 @@
         public short Type;
         public int Asset;
         public uint ModifierTreeId;
-        public byte StartEvent;
+        public int StartEvent;
         public int StartAsset;
         public ushort StartTimer;
-        public byte FailEvent;
+        public int FailEvent;
         public int FailAsset;
-        public byte Flags;
+        public int Flags;
         public short EligibilityWorldStateID;
         public sbyte EligibilityWorldStateValue;
     }
