+++ WoWDeveloperAssistant/DB2/Structures/CriteriaTree.cs
@@ -6,7 +6,7 @@
         public string Description;
         public uint Parent;
         public uint Amount;
-        public sbyte Operator;
+        public int Operator;
         public uint CriteriaID;
         public int OrderIndex;
         public int Flags;
